feat(services): add BaseService and its children services
Add BaseService and its children services: ChangeService, CurrencyService, and RuleService. BaseService contains common methods for all services, while each child service has its own specific methods.

The methods added to ChangeService are: removeRule and addRule. The middleware method in ChangeService adds a new rule to the response of get and removeRule methods.

The methods added to CurrencyService are: get, set, and delete.

The methods added to RuleService are: set and delete. The middleware method in RuleService adds a new rule to the response of set method.